third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/coop-same-origin-allow-popups-document-write-expected.txt

This is a testharness.js-based test.
FAIL coop-same-origin-allow-popups-document-write assert_equals: opener == null expected "true" but got "false"
Harness: the test ran to completion.
